+++ textx/generators.py
@@ -1,4 +1,8 @@
-import click
+try:
+    import click
+except ModuleNotFoundError:
+    raise Exception('textX must be installed with CLI dependencies to use '
+                    'textx command.\npip install textX[cli]')
 import os
 from functools import partial
 from textx.registration import generator